Add available fields in Deluge platform
I've updated the deluge platform to show more available fields to the user, similar to how it is for the PiHole platform. I've pulled the fields directly from the "stats" dictionary. 

These fields are not easy to find on Google. I think this will help most users, so they don't have to use Postman to find all of the fields.

While researching, I also noticed more fields under "result", not under "stats", such as "state", which shows the amount of torrents downloading, seeding, active, paused, etc, which might be a useful statistic?