Update plex.py
correction to the format of the data source for plex. in the beginning of the file it states 'url' instead of 'host' which shows the wrong info in the right info pane and also breaks the data source.

changing it to 'host' fixes the issue, the correct information is shown in the right info pane and the data source works correctly.